Re-organize redux related files.
