+++ meteor-app/imports/ui/redux-store/actions/change_route.js
