+++ meteor-app/imports/ui/redux-store/actions/charts.js
