+++ meteor-app/imports/ui/redux-store/actions/model_inspect_point.js
