+++ meteor-app/imports/ui/redux-store/actions/model_toggle_map.js
