+++ meteor-app/imports/ui/redux-store/actions/open_drawer.js
