+++ meteor-app/imports/ui/redux-store/actions/workspace.js
