+++ meteor-app/imports/ui/redux-store/initial-state.js
