+++ meteor-app/imports/ui/redux-store/reducers/change_route.js
