+++ meteor-app/imports/ui/redux-store/reducers/charts.js
