+++ meteor-app/imports/ui/redux-store/reducers/index.js
