+++ meteor-app/imports/ui/redux-store/reducers/model_inspect_point.js
