+++ meteor-app/imports/ui/redux-store/reducers/model_toggle_map.js
