+++ meteor-app/imports/ui/redux-store/reducers/open_drawer.js
