+++ meteor-app/imports/ui/redux-store/reducers/page_entry.js
