+++ meteor-app/imports/ui/redux-store/reducers/searchpage.js
