+++ meteor-app/imports/ui/redux-store/reducers/workspace.js
